+++ js/data.js
@@ -1,0 +1,80 @@
+const TITLE = [
+  'Квартиры посуточно в городе Токио!',
+  'Красивый вид из окна обеспечен!',
+  'Квартиры с японским интерьером!',
+];
+
+
+const PriceRange = {
+  MIN: 5000,
+  MAX: 10000,
+};
+
+const TYPE = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel'
+];
+
+const RoomsRange =  {
+  MIN: 1,
+  MAX: 3,
+};
+
+const GuestsRange = {
+  MIN: 1,
+  MAX: 5,
+};
+
+const CHECKING = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+const CHECKOUT = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+
+const DESCRIPTION = [
+  'Квартиры с хорошим ремонтом',
+  'Хорошая мебель и просторные комнаты',
+  'В кравтирах проведён быстрый интернет',
+];
+
+const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg,',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+
+const latLocation = {
+  MIN: 35.65000,
+  MAX: 35.70000,
+};
+
+const lngLocation = {
+  MIN: 139.70000,
+  MAX: 139.80000,
+};
+
+
+const AvatarRange = {
+  MIN: 1,
+  MAX: 10,
+};
+
+export {AvatarRange, lngLocation, latLocation , PHOTOS,  FEATURES, CHECKING, PriceRange, DESCRIPTION, CHECKOUT, TYPE, TITLE,  RoomsRange, GuestsRange};
