+++ js/main.js
@@ -1,148 +1,3 @@
-const getRandomNumber = (min, max) => {
-  const rand = min - 0.5 + Math.random() * (max - min + 1);
-  if (min >= max) {
-    return new Error('неверный числовой диапазон');
-  }
-  return Math.round(rand);
-};
-
-
-const getRandomFloatNumber = (min, max, quantityNumber) => {
-  const randomPoint = Math.random() * (max - min) + min;
-  if (min >= max) {
-    return new Error('неверный числовой диапазон');
-  }
-  return randomPoint.toFixed(quantityNumber);
-};
-
-
-const TITLE = [
-  'Квартиры посуточно в городе Токио!',
-  'Красивый вид из окна обеспечен!',
-  'Квартиры с японским интерьером!',
-];
-
-
-const PriceRange = {
-  MIN: 5000,
-  MAX: 10000,
-};
-
-const TYPE = [
-  'palace',
-  'flat',
-  'house',
-  'bungalow',
-  'hotel'
-];
-
-const RoomsRange =  {
-  MIN: 1,
-  MAX: 3,
-};
-
-const GuestsRange = {
-  MIN: 1,
-  MAX: 5,
-};
-
-const CHECKING = [
-  '12:00',
-  '13:00',
-  '14:00',
-];
-
-const CHECKOUT = [
-  '12:00',
-  '13:00',
-  '14:00',
-];
-
-const FEATURES = [
-  'wifi',
-  'dishwasher',
-  'parking',
-  'washer',
-  'elevator',
-  'conditioner',
-];
-
-const DESCRIPTION = [
-  'Квартиры с хорошим ремонтом',
-  'Хорошая мебель и просторные комнаты',
-  'В кравтирах проведён быстрый интернет',
-];
-
-const PHOTOS = [
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg,',
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
-];
-
-const latLocation = {
-  MIN: 35.65000,
-  MAX: 35.70000,
-};
-
-const lngLocation = {
-  MIN: 139.70000,
-  MAX: 139.80000,
-};
-
-
-const AvatarRange = {
-  MIN: 1,
-  MAX: 10,
-};
-
-
-const getArrayRandElement = (array) => {
-  const rand = Math.floor(Math.random() * array.length);
-  return array[rand];
-};
-
-const getRandomArrayLength = (array) => {
-  const newArray = [];
-
-  for (let i = 0; i < getRandomNumber(1, array.length); i++) {
-    newArray.push(array[i]);
-  }
-  return newArray;
-};
-
-
-const getCardGenerate = () => {
-  const latCoordinate = getRandomFloatNumber(latLocation.MIN, latLocation.MAX, 5);
-  const lngCoordinate = getRandomFloatNumber(lngLocation.MIN, lngLocation.MAX, 5);
-  const randomNum = getRandomNumber(AvatarRange.MIN, AvatarRange.MAX);
-
-  return {
-    offer :{
-      title: getArrayRandElement(TITLE),
-      address: `${latCoordinate}, ${lngCoordinate}`,
-      price: getRandomNumber(PriceRange.MIN, PriceRange.MAX),
-      type: getArrayRandElement(TYPE),
-      rooms: getRandomNumber(RoomsRange.MIN, RoomsRange.MAX),
-      guests: getRandomNumber(GuestsRange.MIN, GuestsRange.MAX),
-      checking: getArrayRandElement(CHECKING),
-      checkout: getArrayRandElement(CHECKOUT),
-      features: getRandomArrayLength(FEATURES),
-      description: getArrayRandElement(DESCRIPTION),
-      photos: getRandomArrayLength(PHOTOS),
-    },
-
-    author: {
-      avatar: `img/avatars/user${randomNum === 10 ? '10': `0${randomNum}`}.png`,
-    },
-
-    location: {
-      lat: latCoordinate,
-      lng: lngCoordinate,
-    },
-  };
-};
-
-
-const createCard = () =>  new Array(10).fill(null).map(getCardGenerate);
-
-createCard();
+import './data.js';
+import './util.js';
+import './cardGenerate.js';
